graduate run tasks to plugin
