+++ affectedmoduledetector/src/main/kotlin/com/dropbox/affectedmoduledetector/AffectedModuleDetector.kt
@@ -509,43 +509,4 @@ class AffectedModuleDetectorImpl constructor(
     }
 }
 
-class AffectedModuleConfiguration {
-    /**
-     * Folder to place the log in
-     */
-    var logFolder: String? = null
-
-    /**
-     * Name for the log file
-     */
-    var logFilename: String = "affected_module_detector.log"
-
-    /**
-     * Base directory to use for [pathsAffectingAllModules]
-     */
-    var baseDir: String? = null
-    /**
-     * Files or folders which if changed will trigger all projects to be considered affected
-     */
-    var pathsAffectingAllModules = setOf<String>()
-        set(value) {
-            requireNotNull(baseDir) {
-                "baseDir must be set to use pathsAffectingAllModules"
-            }
-            field = value
-        }
-        get() {
-            field.forEach { path ->
-                require(File(baseDir, path).exists()) {
-                    "Could not find expected path in pathsAffectingAllModules: $path"
-                }
-            }
-            return field
-        }
-
-    companion object {
-        const val name = "affectedModuleDetector"
-    }
-}
-
 val Project.isRoot get() = this == rootProject
