--- sample/buildSrc/src/main/resources/META-INF/gradle-plugins/com.dropbox.affectedtasksplugin.properties
@@ -1,1 +1,0 @@
-implementation-class=com.dropbox.affectedmoduledetector.tasks.AffectedTasksPlugin
